requirements.txt
ag2==0.9.9
alembic==1.16.5
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
asyncer==0.0.8
attrs==25.3.0
autogen==0.9.9
black==25.1.0
blinker==1.9.0
branca==0.8.1
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorlog==6.9.0
contourpy==1.3.3
cycler==0.12.1
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
fastjsonschema==2.21.1
fix-busted-json==0.0.18
flake8==7.3.0
Flask==3.1.1
folium==0.20.0
fonttools==4.59.2
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httplib2==0.30.0
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
jupyter_core==5.8.1
jupytext==1.17.2
kiwisolver==1.4.9
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mccabe==0.7.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
mypy_extensions==1.1.0
narwhals==2.2.0
nbformat==5.10.4
numpy==2.2.6
ollama==0.5.3
openai==1.102.0
opencv-python==4.12.0.88
optuna==4.5.0
packaging==25.0
pandas==2.3.2
pathspec==0.12.1
-e git+ssh://git@github.com/Narendersingh007/perceptron-logic-gates.git@00ef8a81ca61c583914a85df232372820268f389#egg=perceptron_logic_gates
pillow==11.3.0
platformdirs==4.3.8
plotly==6.3.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==6.32.0
pvlib==0.13.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pyflakes==3.4.0
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.8.29
requests==2.32.5
rpds-py==0.27.1
rsa==4.9.1
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
streamlit==1.49.1
streamlit-folium==0.25.1
streamlit-option-menu==0.4.0
tenacity==9.1.2
termcolor==3.1.0
threadpoolctl==3.6.0
tiktoken==0.11.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
xgboost==3.0.4
xyzservices==2025.4.0
